Use `Double` instead of `Float` to represent location values

+++ koleo/src/main/java/xyz/czanik/distanceapp/distance/KoleoStation.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 internal data class KoleoStation(
     @SerializedName("id") val id: Int,
     @SerializedName("name") val name: String,
-    @SerializedName("latitude") val latitude: Float,
-    @SerializedName("longitude") val longitude: Float,
+    @SerializedName("latitude") val latitude: Double,
+    @SerializedName("longitude") val longitude: Double,
     @SerializedName("hits") val hits: Int
 )
